화면 로딩이 느려지면서 전투가 멈추는 현상 fix

+++ content.js
@@ -182,20 +182,29 @@ function battlePageAction() {
 			return;
 		}
 		
-		// 자동사냥 기능을 비활성화한다.
-		const formDom = findBattleButton();
-		if (formDom) {
-			formDom.removeAttribute("name")
-		} else {
-			return;
-		}
+		// 자동사냥 기능을 비활성화한다. DOM이 생성되지 않을 것을 대비해서 1초 딜레이 후, 비활성화
+		const formRemoveWorker = create1000msTimeoutWorker(function() {
+			const formDom = findBattleButton();
+			if (formDom) {
+				formDom.removeAttribute("name")
+			}
+			
+			formRemoveWorker.terminate();
+		});
 		
 		var callCount = 0;
 		const pageLoadTime = new Date();
 		getBattleDuration(function(battleDuration) {
+			var retry = 0;
 			const worker = create50msIntervalWorker(function() {
 				if (!document.querySelector("#reload .hanna")) {
-					worker.terminate();
+					retry += 1;
+					console.log("retry...", retry);
+					// 2초간은 retry 시도해줌
+					if (retry > 40) {
+						console.log("Cannot found DOM");
+						worker.terminate();
+					}
 					return;
 				}
 				const currentTime = new Date();
